Categoria htmls

See merge request pacheco.gonzalez.ivan/erp-ins-cavall-bernat!12

+++ src/main/java/cat/copernic/erpInsCavallBernat/model/Categoria.java
@@ -28,7 +28,7 @@ public class Categoria implements Serializable {
 
     @Id //Indica al sistema que l'atribut idgos és la clau primària de la BBDD
     @GeneratedValue(strategy = GenerationType.IDENTITY) //Indica al sistema com generarem l'id
-    private int id_Categoria;
+    private long id_Categoria;
     /*Validació per comprovar que el nom no està buit. Com a paràmetre li passem el missatge
      *que volem que aparegui.
      */
